#869 split api and db service configuration
